[All] Convert main project to Kotlin
This commit converts the main project to use Kotlin, whilst trying to keep
the codebase mostly the same. The only thing that has changed is that a
BindingsRegistry class has been added. Some things still need to be done,
like a custom setter for BindingsRegistry fields as at the moment, they
can be mutated by anyone even if they were set before (which we don't want).

The version specific modules still need to be converted to Kotlin, but they
will be done sometime in the future.